\name{eventsButton}
\alias{eventsButton}
\title{eventsButton}
\usage{
  eventsButton(inputId, label, events = c("dblclick"), ...)
}
\arguments{
  \item{inputId}{the id of the function.}

  \item{label}{The displayed text}

  \item{events}{A sequence of events that will invalidate
  the button
  \url{http://getbootstrap.com/2.3.2/base-css.html#icons}}
}
\description{
  A button that responds to events click, doubleclick,
  hover etc. It's built on top of actionButton and
  eventInput
}

