\name{hideshinyalert}
\alias{hideshinyalert}
\title{hideshinyalert}
\usage{
hideshinyalert(session, id)
}
\arguments{
  \item{id}{Specifies the alert id that will be used to
  access the}
}
\description{
Used in server.R. Hide an alert placed in ui.R with
shinyalert
}
\seealso{
shinyalert, showshinyalert

Other ShinySky.elements: \code{\link{actionButton}};
\code{\link{select2Input}}; \code{\link{shinyalert}};
\code{\link{textInput.typeahead}}
}

