% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotable.r
\name{hot.to.df}
\alias{hot.to.df}
\title{hot.to.df}
\usage{
hot.to.df(b)
}
\arguments{
\item{b}{The input$hotable_id value.}
}
\description{
Converts the table data passed from the client-side into a data.frame
}
