\name{includeSelect2}
\alias{includeSelect2}
\title{includeSelect2}
\usage{
includeSelect2()
}
\value{
the HTML to include the Javascripot and CSS of select2.js
}
\description{
Some code to include css and js for includeSelect2. This
provide so that the use can have access to select2 even if
they only use it in the dynamic UI
}

