% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jstree.R
\name{jstree.obj}
\alias{jstree.obj}
\title{jstree.obj}
\usage{
jstree.obj(x, addDepthClass = TRUE, addIndex = TRUE, depth = 0,
  index = "1")
}
\arguments{
\item{x}{The id used to refer to the jstree}

\item{addDepthClass}{Add a class attribute "treedepth_n" to each node, where n is the tree depth (depth) of each node, starting with top depth = 1.  Useful with conditionalPanel, css and jQuery.}

\item{addIndex}{Add an attrbute "index" to each node, a string containing the path from the top, in concatenated ind values.}

\item{depth}{The current tree depth for this node, to be passed into the recursive calls. Default = 0 (top depth). Not intended for the user; used only in the recursion.}

\item{index}{A string containing the path from the top, in concatenated ind values.  Not intended for the user; used only in the recursion.}
}
\description{
Creates a jstree object to be passed to jstree for creating a tree view
}

