% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jstree.R
\name{jstree}
\alias{jstree}
\title{jstree}
\usage{
jstree(id, tree.list)
}
\arguments{
\item{id}{The id used to refer to the jstree}

\item{tree.list}{The tree list create dusing jstree.obj}
}
\description{
Creates a jstree
}

