% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotable.r
\name{renderHotable}
\alias{renderHotable}
\title{renderHotable}
\usage{
renderHotable(expr, env = parent.frame(), quoted = FALSE,
  options = NULL, readOnly = TRUE)
}
\arguments{
\item{expr}{The computation that leads to an output}

\item{env}{The R environment in which to create the dataset}

\item{quoted}{Pass to the exprToFunction}

\item{options}{Pass to the exprToFunction}

\item{readOnly}{A vector of TRUE/FALSE values to indicate which of the 
columns should be read-only.}
}
\description{
Renders the hotable.
}
\examples{

\dontrun{
# these example will only run inside a shiny app

### ui.R
#' hotable("hotable1")
#This will create a handsontable which you can output using 
output$hotable1 <- renderHotable({...})

### server.R
something <- reactive({
    hot.to.df(input$hotable1) # this will convert your input into a data.frame
})
}

}
