% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-shiny-sky-example.R
\name{run.shinysky.example}
\alias{run.shinysky.example}
\title{Run shinysky example shinyapp}
\usage{
run.shinysky.example()
}
\description{
Run shinysky example shinyapp
}
\examples{

# simply run it to show a Shiny app

if(interactive()) {
  shiny::run.shinysky.example()
}


}
