% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny-alert.R
\name{shinyalert}
\alias{shinyalert}
\title{shinyalert}
\usage{
shinyalert(id, click.hide = TRUE, auto.close.after = NULL)
}
\arguments{
\item{id}{Specifies the alert id that will be used to access the}

\item{click.hide}{If TRUE then clicking on the alert will hide it. Defaults to TRUE}

\item{auto.close.after}{After this many seconds auto close the alert}
}
\description{
Creates an shinyalert
}
\seealso{
shinyalert

Other ShinySky.elements: \code{\link{actionButton}},
  \code{\link{eventsButton}}; \code{\link{hideshinyalert}};
  \code{\link{select2Input}}; \code{\link{showshinyalert}};
  \code{\link{textInput.typeahead}},
  \code{\link{updateTextInput.typeahead}}
}

