\name{shinysky-package}
\alias{shinysky-package}
\alias{shinysky}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ShinySky is set of UI components/widgets
}
\description{
ShinySky is set of UI components/widgets
}
\details{
  \tabular{ll}{
  Package: \tab shinysky\cr
  Type: \tab Package\cr
  Version: \tab 0.1.0\cr
  Date: \tab 2013-11-15\cr
  License: \tab MIT\cr
  }
ShinySky Components
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
AnalytixWare
}
\note{
%%  ~~further notes~~
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ package }

