\name{showShinyAlert}
\alias{showShinyAlert}
\title{showShinyAlert}
\usage{
  showShinyAlert(id, HTMLtext, session,
    styleclass = "success")
}
\arguments{
  \item{id}{Specifies the alert id that will be used to
  access the}

  \item{HTMLtext}{The HTML you want displayed in the text}

  \item{styleclass}{The styling class of the alert--options
  are primary, info, success ,warning, ,danger ,inverse
  ,link}
}
\description{
  Show an alert
}
\seealso{
  Other shinysky elements: \code{\link{shinyalert}}
}

