% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateSelect2Input}
\alias{updateSelect2Input}
\title{updateSelect2Input}
\usage{
updateSelect2Input(session, inputId, label, choices = NULL, selected = NULL)
}
\arguments{
\item{session}{The session}

\item{inputId}{The id of the component}

\item{label}{The text label}

\item{choices}{The possible choices}

\item{selected}{The selected.}
}
\description{
Update select2Input
}

