% Generated by roxygen2 (4.0.2): do not edit by hand
\name{actionButton}
\alias{actionButton}
\alias{eventsButton}
\title{Buttons including Action Buttons and Event Buttons}
\usage{
actionButton(inputId, label, styleclass = "", size = "", block = F,
  icon = NULL, icon.library = c("bootstrap", "font awesome"),
  css.class = "", ...)

eventsButton(..., events = c("dblclick"))
}
\arguments{
\item{inputId}{Specifies the input slot that will be used to access the
value.}

\item{label}{The contents of the button--usually a text label, but you could
also use any other HTML, like an image.}

\item{styleclass}{The Bootstrap styling class of the button--options are
primary, info, success, warning, danger, inverse, link or blank}

\item{size}{The size of the button--options are large, small, mini}

\item{block}{Whehter the button should fill the block}

\item{icon}{Display an icon for the button}

\item{icon.library}{Specify an icon set to use
\url{http://www.fontawesome.io/icons} or
\url{http://getbootstrap.com/2.3.2/base-css.html#icons}}

\item{css.class}{Any additional CSS class one wishes to add to the action
button}

\item{...}{Other argument to feed into shiny::actionButton}

\item{events}{A sequence of events that will invalidate the button
\url{http://getbootstrap.com/2.3.2/base-css.html#icons}}
}
\description{
Creates an action button whose value is initially zero, and increments by one
each time it is pressed.

A button that responds to events click, doubleclick, hover etc. It's built on
top of actionButton and eventInput
}
\seealso{
Other ShinySky.elements: \code{\link{hideshinyalert}};
  \code{\link{select2Input}}; \code{\link{shinyalert}};
  \code{\link{showshinyalert}};
  \code{\link{textInput.typeahead}},
  \code{\link{updateTextInput.typeahead}}
}

