% Generated by roxygen2 (4.0.2): do not edit by hand
\name{busyIndicator}
\alias{busyIndicator}
\title{busyIndicator}
\usage{
busyIndicator(text = "Calculation in progress..",
  img = "shinysky/busyIndicator/ajaxloaderq.gif", wait = 1000)
}
\arguments{
\item{text}{The text to show}

\item{img}{An anitmated gif}

\item{wait}{The amount of time to wait before showing the busy indicator. The
  default is 1000 which is 1 second.}
}
\description{
A busy indicator
}

