% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hideshinyalert}
\alias{hideshinyalert}
\title{hideshinyalert}
\usage{
hideshinyalert(session, id)
}
\arguments{
\item{id}{Specifies the alert id that will be used to access the}
}
\description{
Used in server.R. Show an alert placed in ui.R with shinyalert
}
\seealso{
shinyalert

Other ShinySky.elements: \code{\link{actionButton}};
  \code{\link{select2Input}}; \code{\link{shinyalert}};
  \code{\link{showshinyalert}};
  \code{\link{textInput.typeahead}}
}

