% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny-alert.R
\name{hideshinyalert}
\alias{hideshinyalert}
\title{hideshinyalert}
\usage{
hideshinyalert(session, id)
}
\arguments{
\item{session}{The Shiny sessoin specified in shinyServer(function(input, output, session) {...}).}

\item{id}{Specifies the alert id that will be used to access the}
}
\description{
Used in server.R. Show an alert placed in ui.R with shinyalert
}
\seealso{
shinyalert

Other ShinySky.elements: \code{\link{actionButton}},
  \code{\link{eventsButton}}; \code{\link{select2Input}};
  \code{\link{shinyalert}}; \code{\link{showshinyalert}};
  \code{\link{textInput.typeahead}},
  \code{\link{updateTextInput.typeahead}}
}

