% Generated by roxygen2 (4.0.2): do not edit by hand
\name{includeSelect2}
\alias{includeSelect2}
\title{includeSelect2}
\usage{
includeSelect2(drag.and.drop = FALSE)
}
\value{
the HTML to include the Javascripot and CSS of select2.js
}
\description{
Some code to include css and js for includeSelect2. This is provide so that
the user can have access to select2 even if they only use it in the dynamic
UI
}

