% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hotable.r
\name{renderHotable}
\alias{renderHotable}
\title{renderHotable}
\usage{
renderHotable(expr, env = parent.frame(), quoted = FALSE, options = NULL,
  readOnly = TRUE)
}
\arguments{
\item{expr}{The computation that leads to an output}

\item{env}{The R environment in which to create the dataset}

\item{quoted}{Pass to the exprToFunction}

\item{options}{Pass to the exprToFunction}

\item{readOnly}{A vector of TRUE/FALSE values to indicate which of the
  columns should be readonly.}
}
\description{
Renders the hotable.
}

