\name{select2Input}
\alias{select2Input}
\title{select2 Input}
\usage{
select2Input(inputId, ...)
}
\arguments{
  \item{inputId}{Specifies the input slot that will be used
  to access the value.}
}
\description{
The select2 input
}
\seealso{
Other ShinySky.elements: \code{\link{actionButton}};
\code{\link{shinyalert}}; \code{\link{showshinyalert}};
\code{\link{textInput.typeahead}}
}

