% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select2Input}
\alias{select2Input}
\title{select2 Input}
\usage{
select2Input(inputId, label, choices = NULL, selected = NULL,
  type = c("input", "select"), drag.and.drop = FALSE, ...)
}
\arguments{
\item{inputId}{Input variable to assign the control's value to}

\item{label}{Display label for the control}

\item{choices}{List of values to select from. If elements of the list are
named then that name rather than the value is displayed to the user.}

\item{selected}{Name of initially selected item (or multiple names if
\code{multiple = TRUE}). If not specified then defaults to the first item
for single-select lists and no items for multiple select lists.}

\item{type}{The two valid options are input and select. The input type is
similar to a text input and it can remember the order of selection. The
select is based on the <select> tags in HTML and is not ordinal hence the
order in which you specify the selection is not being taken into account,
however it's input parameters are compatible with selectInput and you can
use updateSelectInput to update its contents.}

\item{drag.and.drop}{Allow drag and drop or not}

\item{...}{Other arguments to pass into shiny::selectInput}
}
\value{
A select list control that can be added to a UI definition.#'
}
\description{
The select2 input
}
\seealso{
Other ShinySky.elements: \code{\link{actionButton}},
  \code{\link{eventsButton}}; \code{\link{hideshinyalert}};
  \code{\link{shinyalert}}; \code{\link{showshinyalert}};
  \code{\link{textInput.typeahead}},
  \code{\link{updateTextInput.typeahead}}
}

