% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-alert.R
\name{shinyalert}
\alias{shinyalert}
\alias{showshinyalert}
\alias{hideshinyalert}
\title{shinyalert}
\usage{
shinyalert(id, click.hide = TRUE, auto.close.after = NULL)

showshinyalert(session, id, HTMLtext, styleclass = "success")

hideshinyalert(session, id)
}
\arguments{
\item{id}{Specifies the alert id that will be used to access the}

\item{click.hide}{If TRUE then clicking on the alert will hide it. Defaults to TRUE}

\item{auto.close.after}{After this many seconds auto close the alert}

\item{session}{The Shiny session specified in shinyServer(function(input, output, session) {...}).}

\item{HTMLtext}{The HTML you want displayed in the text}

\item{styleclass}{The styling class of the alert--options are primary, info, 
success ,warning, ,danger ,inverse ,link}
}
\description{
Creates an shinyalert. The show and hide functions are used in server.R. Show
an alert placed in ui.R with shinyalert

Used in server.R. Show an alert placed in ui.R with shinyalert
}
\examples{

# ui.R

shinyalert("id", click.hide = TRUE, auto.close.after = NULL)

# close after one second
# and clicking doesn't hide it
shinyalert("id", click.hide = FALSE, auto.close.after = 1)

}
\seealso{
shinyalert

shinyalert

shinyalert

Other ShinySky elements: \code{\link{actionButton}},
  \code{\link{select2Input}},
  \code{\link{textInput.typeahead}}

Other ShinySky elements: \code{\link{actionButton}},
  \code{\link{select2Input}},
  \code{\link{textInput.typeahead}}

Other ShinySky elements: \code{\link{actionButton}},
  \code{\link{select2Input}},
  \code{\link{textInput.typeahead}}
}
\concept{ShinySky elements}
