\name{shinyalert}
\alias{shinyalert}
\title{shinyalert}
\usage{
  shinyalert(id, click.hide = TRUE)
}
\arguments{
  \item{id}{Specifies the alert id that will be used to
  access the}

  \item{click.hide}{If TRUE then clicking on the alert will
  hide it. Defaults to TRUE}
}
\description{
  Creates an shinyalert
}
\seealso{
  shinyalert

  Other ShinySky.elements: \code{\link{actionButton}},
  \code{\link{select2Input}}, \code{\link{showshinyalert}},
  \code{\link{textInput.typeahead}}
}

