\name{shinyalert}
\alias{shinyalert}
\title{shinyalert}
\usage{
  shinyalert(id)
}
\arguments{
  \item{id}{Specifies the alert id that will be used to
  access the}
}
\description{
  Creates an shinyalert
}
\seealso{
  shinyalert

  Other ShinySky.elements: \code{\link{actionButton}},
  \code{\link{select2Input}}, \code{\link{showshinyalert}},
  \code{\link{textInput.typeahead}}
}

