% Generated by roxygen2 (4.0.2): do not edit by hand
\name{textInput.typeahead}
\alias{textInput.typeahead}
\alias{updateTextInput.typeahead}
\title{textinput.typeahead}
\usage{
textInput.typeahead(id, placeholder, local, valueKey, tokens, template,
  limit = 20)

updateTextInput.typeahead(session, id, dataset, valueKey, tokens, template,
  limit = 20, placeholder = "")
}
\arguments{
\item{id}{The id of the field}

\item{placeholder}{The placeholder text. It is shown in the text input before
any input has occured}

\item{local}{A data.frame containing all the information you want to make
available to typeahead to display in the dropdown. You can use any of the
variable in the dataset via the template argument}

\item{valueKey}{The element that acts as the value key in the dataset usually
an unique identifier of the row}

\item{tokens}{A list whose length equal to nrow(local) where each element is
array of string tokens. Typing the tokens will select the correponding rows}

\item{template}{A hogan template on how to show the content e.g. '<p>
  {{name}} </p>' where name is one of the variables in local}

\item{limit}{An integer of the upper limit on how many hits to show in the
typeahead dropdown}

\item{session}{The Shiny sessoin specified in shinyServer(function(input, output, session) {...}).}

\item{dataset}{The data.frame to render}
}
\description{
A simple typeahead text input component

Update textInput.typeahead
}
\seealso{
Other ShinySky.elements: \code{\link{actionButton}},
  \code{\link{eventsButton}}; \code{\link{hideshinyalert}};
  \code{\link{select2Input}}; \code{\link{shinyalert}};
  \code{\link{showshinyalert}}
}

